début passage interactions DB dans module dédié.
